requirements.txt
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.15,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-chroma>=0.2.0,<0.3.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-experimental>=0.3.0,<0.4.0
langchain-cohere>=0.3.0,<0.4.0
langchain-voyageai
google-genai
python-dotenv
rank_bm25>=0.2.2